src/main/java/com/example/vmo1/service/impl/ImageServiceImpl.java
package com.example.vmo1.service.impl;

import com.cloudinary.Cloudinary;
import com.cloudinary.utils.ObjectUtils;
import com.example.vmo1.model.entity.Image;
import com.example.vmo1.repository.ImageRepository;
import com.example.vmo1.service.ImageService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;
import org.springframework.web.multipart.MultipartFile;

import java.io.*;
import java.util.HashMap;
import java.util.List;
import java.util.Map;

@Service
public class ImageServiceImpl implements ImageService {

    @Autowired
    private ImageRepository imageRepository;

    @Override
    public List<Image> saveImage(List<Image> image){
        return imageRepository.saveAll(image);
    }





}
